+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - `Strm` output has an additional option `strm_props`. These props are written to the strm file.
 If `kodi_style` set to `true` the property `#KODIPROP:seekable=true|false` is added. If `strm_props` is not given `#KODIPROP:inputstream=inputstream.ffmpeg`, `"#KODIPROP:http-reconnect=true` are set too for `kody_style`.
 - Fixed xtream affix-processed output.
+- `log_sanitize_sensitive_info`  added to `config.yml`. Default is `true`.
 
 # 2.1.1 (2025-01-19)
 - added new path `/status` which is an alias to `healthcheck`